renamed pages, reworked homepage, added button to take you to motivation page
