+++ ct_gen/src/pages/page_4.py
@@ -7,11 +7,12 @@ from ct_gen.src.modules.image_functions import *
 
 
 def display_page_4():
-    image_path = "ct_gen/data/images/culprits"
-    step_title = "Step 2"
-    title = "The Conspirators"
-    info = "Who’s behind it? Every conspiracy theory needs a sinister group of scheming culprits. Pick one from the selection below."
-    sheet_name = "culprits"
-    instruction = "Select a culprit"
-    create_image_selection_view(image_path, step_title, title, info, sheet_name, instruction)
     scroll_up()
+    image_path = "ct_gen/data/images/motives"
+    step_title = "Step 3"
+    title = "The Motive"
+    info = "What's their endgame? Every conspiracy theory has a motive. Select one of the options below."
+    sheet_name = "motives"
+    instruction = "Select a motive"
+    create_image_selection_view(image_path, step_title, title, info, sheet_name, instruction)
+
