+++ ct_gen/src/pages/page_6.py
@@ -1,33 +1,44 @@
 import streamlit as st
+import datetime
+import gspread
+#from gsheetsdb import connect
+import shillelagh
+import sqlite3
+import base64
+import newspaper
+from oauth2client.service_account import ServiceAccountCredentials
+from oauth2client import service_account
 import os
 import openai
 import pandas as pd
+import random
+import requests
+#from streamlit_extras.badges import badge
+import time
+import sys
+import webbrowser
 from openai import OpenAI
 import toml
 from ct_gen.src.modules.image_functions import display_list_of_images
 from ct_gen.src.modules.rating_buttons import add_rating_buttons
 from ct_gen.src.modules.google_sheets_api import insert_row_to_sheet, connect_to_google_sheets_data
 from ct_gen.src.modules.initialize_session_state import initalize_session_state_dict
-#from ct_gen.src.modules.pdf_download import add_pdf_button
-import streamlit.components.v1 as components
-from ct_gen.src.modules.scroll_up import scroll_up
+from ct_gen.src.modules.pdf_download import add_pdf_button
 
-def create_prompt():
+
+def create_explanation_prompt():
     selected_article_content = st.session_state["news_summary"]
     culprits = st.session_state["culprits_name"]
     culprits_info = st.session_state["culprits_summary"]
     motive = st.session_state["motives_name"]
     motive_info = st.session_state["motives_summary"]
+    
+    ct = st.session_state["conspiracy_theory"]
+    prompt = st.session_state["prompt"]
         
-    prompt = f"""Write a convicing conspiracy theory by turning the following news story into a conspiracy theory: {selected_article_content}
-        The conspirator(s) of your story are: {culprits} ({culprits_info}).
-        The motive of these conspirators: {motive} ({motive_info}).
-        You construct the conspiracy by following these steps:
-        You find some suspicious loopholes, puzzling details and anomalies in the official story. You 'just ask questions' in the style of conspiracy theorists.
-        You fabricate some 'evidence' that {selected_article_content} is a cover-up of {culprits} trying to achieve {motive}. You 'connect the dots' in the style of conspiracy theorists, using available information about {culprits}.
-        You anticipate counterarguments against the conspiracy theory by arguing that missing evidence and counterevidence is in fact part of the plot. Make sure to make the conspiracy theory immune against criticism
-        You discredit people who are sceptical of the conspiracy theory by suggesting they are gullible dupes or patsies complicit in the conspiracy
-        Write a convicing story starting with a catchy title. Simplify the story to a reading level of grade ten and shorten the story. Everything must be formated in markdown."""
+    prompt = f"""You have created the following conspiracy theory
+        {ct} by following the recipe {prompt}.
+        Reflect on how you constructed this theory and explain it in simple terms."""
 
     return prompt
 
@@ -41,7 +52,7 @@ def load_secrets():
 
 # Generate CT function
 @st.cache_data()
-def generate_conspiracy_theory(prompt, _client):
+def generate_explanation(prompt, _client):
         
     res_box = st.empty()
     report = []
@@ -60,75 +71,46 @@ def generate_conspiracy_theory(prompt, _client):
             result = "".join(report).strip()
             res_box.markdown(f'{result}') 
             
-    st.session_state["conspiracy_theory"] = "".join(report).strip()
-
-def create_twitter_button():
-    post_text = f"I've just made my own conspiracy theory with the Conspiracy Generator, a new educational tool!\
-    \nHere it is:\n\
-    - Official story: {st.session_state['news_name']}\n\
-    - Culprits: {st.session_state['culprits_name']}\n\
-    - Motive: {st.session_state['motives_name']}\n\
-    Make your own conspiracy here:"
-
-    components.html(
-    f"""
-        <a href="https://twitter.com/share?ref_src=twsrc%5Etfw" class="twitter-share-button" 
-        data-text={post_text} 
-        data-url="https://conspiracy-generator.streamlit.app/"
-        data-show-count="false">
-        data-size="Large" 
-        data-hashtags="streamlit,conspiracy"
-        Tweet
-        </a>
-        <script async src="https://platform.twitter.com/widgets.js" charset="utf-8"></script>
-    """
-    )
-
-  
+    st.session_state["conspiracy_explanation"] = "".join(report).strip()
+    
 # Display page
 def display_page_6():
+    sheet = connect_to_google_sheets_data()
     #initalize_session_state_dict()
-    step_title = "Step 4"
-    title = "Your Conspiracy Theory"
-    info = "See how your selection of culprits and motives turns a simple news story into a conspiracy theory."
+    step_title = "Step 5"
+    title = "Explanation"
+    info = "See how your conspiracy theory got constructed."
     
-    ct_sheet = connect_to_google_sheets_data("generated_ct")
-    ratings_sheet = connect_to_google_sheets_data("ratings")
+    
     st.markdown(f"<h3 style='text-align: center;'>{step_title}</h3>", unsafe_allow_html=True)
     st.markdown(f"<h1 style='text-align: center;'>{title}</h1>", unsafe_allow_html=True)
     st.info(info)
-    scroll_up()
     
     # Load the secrets at the start of the app
     secrets = load_secrets()
     client = OpenAI(api_key=secrets["openai"]["api_key"])
     images = [st.session_state["news_img"], st.session_state["culprits_img"], st.session_state["motives_img"]]
-    captions = ["STORY:\n\n" + st.session_state["news_caption"], "CULPRIT:\n\n" + st.session_state["culprits_caption"], "MOTIVE:\n\n" + st.session_state["motives_caption"]]
+    captions = [st.session_state["news_caption"], st.session_state["culprits_caption"], st.session_state["motives_caption"]]
+    
     display_list_of_images(images, captions)
     
-    st.session_state["prompt"] = create_prompt()
+    st.session_state["explanation_prompt"] = create_explanation_prompt()
    
     st.divider()
+    #generation_button = st.button("Generate your theory!")
+    #if generation_button:
     
-    generate_conspiracy_theory(st.session_state["prompt"], client)
+    generate_explanation(st.session_state["prompt"], client)
     
-    if st.session_state["ct_saved"] == False:
-        
-        ct_row = [
-        st.session_state["news_name"],
-        st.session_state["news_summary"],
-        st.session_state["culprits_name"],
-        st.session_state["culprits_summary"],
-        st.session_state["motives_name"],
-        st.session_state["motives_summary"],
+    row = [
         st.session_state["prompt"],
-        st.session_state["conspiracy_theory"]
-        ]
-        insert_row_to_sheet(ct_sheet, ct_row)
-        st.session_state["ct_saved"] = True
+        st.session_state["conspiracy_theory"],
+        st.session_state["conspiracy_explanation"],
+    ]
+    if st.session_state["conspiracy_explanation"] == "":
+        insert_row_to_sheet(sheet, "generated_ct", row)
     
-    add_rating_buttons(ct_sheet, ratings_sheet)
-    create_twitter_button()
+    #add_rating_buttons(sheet)
     
     # if st.session_state["conspiracy_theory"] != "":
     #     add_pdf_button(st.session_state["conspiracy_theory"])
